Combine all Task4 classes into single file
