--- CmsApplication/Repositories/IAuthorRepository.cs
@@ -1,14 +1,0 @@
-﻿using CmsApplication.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CmsApplication.Repositories
-{
-    public interface IAuthorRepository
-    {
-        Task<List<Author>> GetAll();
-        Task Update(Author entity);
-    }
-}
