+++ InterviewSamples.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InterviewSamples", "InterviewSamples\InterviewSamples.csproj", "{D5791543-CC5E-4734-AACB-22A2BDF885E2}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CmsApplication", "CmsApplication\CmsApplication.csproj", "{449D683C-02FF-45F7-9FB0-46719501E5D8}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{D5791543-CC5E-4734-AACB-22A2BDF885E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D5791543-CC5E-4734-AACB-22A2BDF885E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D5791543-CC5E-4734-AACB-22A2BDF885E2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{449D683C-02FF-45F7-9FB0-46719501E5D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{449D683C-02FF-45F7-9FB0-46719501E5D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{449D683C-02FF-45F7-9FB0-46719501E5D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{449D683C-02FF-45F7-9FB0-46719501E5D8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
