+++ InterviewSamples/Task4.cs
@@ -1,14 +1,61 @@
-﻿using CmsApplication.Entities;
-using CmsApplication.Repositories;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using System.Linq;
 
-namespace CmsApplication
+namespace InterviewSamples
 {
-	public class PostManager : InMemoryPostRepository
+	public class Post
+	{
+		public Guid Id { get; set; }
+		public string Title { get; set; }
+		///Should it be visible to public
+		public bool IsPublicVisible { get; set; }
+		///When it shall become visible to public
+		public DateTime PublishTime { get; set; }
+		public int Category { get; set; }
+		public Guid AuthorId { get; set; }
+	}
+
+	public class Author
+	{
+		public Guid id { get; set; }
+		public bool isEmailVerified { get; set; }
+	}
+
+	public class PostRepository
+	{
+		public async Task<List<Post>> SelectAll()
+		{
+			//will be a db request here in future
+			return new List<Post>();
+		}
+
+		public Task Update(Post entity)
+		{
+			//will update entity in db here in future
+			return Task.CompletedTask;
+		}
+	}
+
+	public class AuthorRepository
+	{
+		public async Task<List<Author>> GetAll()
+		{
+			//will be a db request here in future
+			return new List<Author>();
+		}
+
+		public Task Update(Author entity)
+		{
+			//will update entity in db here in future
+			return Task.CompletedTask;
+		}
+	}
+
+
+	public class PostManager : PostRepository
 	{
 		public async Task<List<Post>> SelectPosts()
 		{
@@ -54,7 +101,7 @@ namespace CmsApplication
 				}
 			}
 
-			IAuthorRepository authorRepository = new InMemoryAuthorRepository();
+			AuthorRepository authorRepository = new AuthorRepository();
 			var authors = await authorRepository.GetAll();
 
 			return result.Join(authors,
